Use initial state in filter
